extract stubs from test file
