+++ Tests/AuthTests/AuthRequstSubscriberTests.swift
@@ -4,7 +4,6 @@ import XCTest
 import WalletConnectUtils
 @testable import WalletConnectKMS
 @testable import TestingUtils
-import JSONRPC
 
 class AuthRequstSubscriberTests: XCTestCase {
     var networkingInteractor: NetworkingInteractorMock!
@@ -38,23 +37,5 @@ class AuthRequstSubscriberTests: XCTestCase {
         wait(for: [messageExpectation], timeout: defaultTimeout)
         XCTAssertEqual(message, expectedMessage)
         XCTAssertEqual(messageId, expectedRequestId)
-    }
-}
-
-extension RequestSubscriptionPayload {
-    static func stub(id: Int64) -> RequestSubscriptionPayload {
-        let appMetadata = AppMetadata(name: "", description: "", url: "", icons: [])
-        let requester = AuthRequestParams.Requester(publicKey: "", metadata: appMetadata)
-        let issueAt = ISO8601DateFormatter().string(from: Date())
-        let payload = AuthPayload(requestParams: RequestParams.stub(), iat: issueAt)
-        let params = AuthRequestParams(requester: requester, payloadParams: payload)
-        let request = RPCRequest(method: "wc_authRequest", params: params, id: id)
-        return RequestSubscriptionPayload(topic: "", request: request)
-    }
-}
-
-extension RequestParams {
-    static func stub() -> RequestParams {
-        return RequestParams(domain: "", chainId: "", nonce: "", aud: "", nbf: nil, exp: nil, statement: nil, requestId: nil, resources: nil)
     }
 }
